Auth e criptografia

+++ app.py
@@ -1,11 +1,14 @@
+import bcrypt
 from flask import Flask, request, jsonify
 from models.user import User
 from config.database import db
 from flask_login import LoginManager, login_user, current_user, logout_user, login_required
 
+
 app = Flask(__name__)
 app.config['SECRET_KEY'] = "your_secret_key"
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///database.db'
+#app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///database.db'
+app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql+pymysql://root:admin123@127.0.0.1:3306/flaskcrud'
 
 login_manager = LoginManager()
 db.init_app(app)
@@ -25,7 +28,7 @@ def login():
 
     if username and password:
         user = User.query.filter_by(username=username).first()
-        if user and user.password == password:
+        if user and bcrypt.checkpw(str.encode(password), str.encode(user.password)):
             login_user(user)
             print(current_user.is_authenticated)
             return jsonify({'message':'Login Successful'}), 200
@@ -47,10 +50,12 @@ def create_user():
     username = data.get('username')
     password = data.get('password')
 
+    hashdpassword = bcrypt.hashpw(str.encode(password), bcrypt.gensalt());
+
     if username and password:
         if User.query.filter_by(username=username).first():
             return jsonify({'message':'Username already exists'}), 409
-        new_user = User(username=username, password=password)
+        new_user = User(username=username, password=hashdpassword)
         db.session.add(new_user)
         db.session.commit()
         return jsonify({'message':'User created successfully'}), 201
@@ -72,10 +77,16 @@ def update_user(id_user):
     if user:
         username = data.get('username')
         password = data.get('password')
+
+
+        if user.id != current_user().id and current_user.role == 'user':
+            return jsonify({'message':'Unauthorized to update this user'}), 403
+
         if username and user.id != current_user().id:
             user.username = username
         if password:
-            user.password = password
+            hashdpassword = bcrypt.hashpw(str.encode(password), bcrypt.gensalt());
+            user.password = hashdpassword
         db.session.commit()
         return jsonify({'message':'User updated successfully'}), 200
     return jsonify({'message':'User not found'}), 404
@@ -95,6 +106,10 @@ def delete_user(id_user):
 
     if id_user==current_user.id:
         return jsonify({'message':'You cannot delete your own account'}), 403
+
+    if current_user.role == 'user':
+        return jsonify({'message':'Unauthorized to delete this user'}), 403
+
 
     if user:
         db.session.delete(user)
